fix: NextAuth 설정 빌드 오류 수정
