+++ src/app/layout.tsx
@@ -11,7 +11,7 @@ import { Inter } from "next/font/google";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "FLAWDETECTOR",
   description: "Generated by create next app",
 };
